Add net45 framework as a target (#56)
Problem: The new regex match timeout feature is enabled only for the
compatible framework. It appears in net45 but the latest net framework
in the project is net40. This means that this feature is not enabled for
all net framework.

Solution: Add net45 as a target framework and enable the feature for
this framework by defining a preprocessor symbol for this framework.

Issue: issues/55